+++ tests/settings.py
@@ -113,6 +113,7 @@ FLOW_EXPRESSION_ENGINES = [
 ]
 FLOW_EXECUTION_ENGINES = [
     'resolwe.flow.execution_engines.bash',
+    'resolwe.flow.execution_engines.workflow',
 ]
 
 FLOW_PROCESSES_FINDERS = (